Do not attempt to build HostLink sources when compiling the Mothership.
As of Tinsel 0.6, the source files for different Tinsel versions are different. Rather than attempting to maintain a list for each version, we defer to the people installing/building Tinsel on this platform. That's either the system Tinsel (currently in /local/tinsel), or the Tinsel in the Orchestrator dependency tarball.