Add some logic to "load" placement from a file.
This is really hacky, and has been implemented so that I can load placement
configurations for simulations without holding the hostlink on a box. Don't use
this unless you really know what you're doing, because it has no error handling
and has been written in a rush on a Friday evening by tired Mark.

It's on a separate branch for a reason. You have been warned.