Bugfix for issue #153
This change addresses the edge case where a Mothership does not exist for a box
in the hardware model (the user is warned...), and the user does not place any
devices on that box. When deployed, the task causes the Orchestrator to
segfault, because there is no entry in the procmap for the Mothership mapped to
the box (because there *is no* mothership mapped to the box!).

The fix is simple - introduce a null comm-rank value, which causes the
deployment to error out gracefully if it is not overridden. In normal
operation, the value is overridden when a device is found on a box *with* a
corresponding Mothership.