Fix use of uninitialised value (valgrind).
Each time OrchBase is issued a load command, the CmLoad instance creates XValid
if it doesn't exist, passing in a file pointer as it does so. Since the file is
closed at the end of every command (Root.cpp:473), the file pointer needs to be
updated with the new file at the beginning of each command.